sdkv2/ch12/cdk/requirements.txt
-e .
aws_cdk.aws_sagemaker
